Fix Cell pos types to be size_t instead of int

+++ include/universe.hpp
@@ -16,7 +16,7 @@ class Universe {
         virtual bool isCellAlive(size_t row, size_t col) = 0;
         virtual void makeCellAlive(size_t row, size_t col) = 0;
         virtual void makeCellDead(size_t row, size_t col) = 0;
-        virtual std::vector<std::pair<int, int>> getAliveCellsPos() = 0;
+        virtual std::vector<std::pair<size_t, size_t>> getAliveCellsPos() = 0;
         size_t rowCount() const { return m_rows; }
         size_t colCount() const { return m_cols; }
         virtual ~Universe() {};
@@ -33,7 +33,7 @@ class DenseUniverse: public Universe {
         bool isCellAlive(size_t row, size_t col) override;
         void makeCellAlive(size_t row, size_t col) override;
         void makeCellDead(size_t row, size_t col) override;
-        std::vector<std::pair<int, int>> getAliveCellsPos() override;
+        std::vector<std::pair<size_t, size_t>> getAliveCellsPos() override;
     private:
         std::vector<Cell*> getNeighbors(Cell* cell);
         std::vector<std::vector<std::unique_ptr<Cell>>> m_cell_grid;
@@ -47,7 +47,7 @@ class SparseUniverse: public Universe {
         bool isCellAlive(size_t row, size_t col) override;
         void makeCellAlive(size_t row, size_t col) override;
         void makeCellDead(size_t row, size_t col) override;
-        std::vector<std::pair<int, int>> getAliveCellsPos() override;
+        std::vector<std::pair<size_t, size_t>> getAliveCellsPos() override;
     private:
         std::set<std::unique_ptr<Cell>>::iterator findAliveCellByPos(size_t row, size_t col);
         std::set<std::unique_ptr<Cell>> m_alive_cells;
