+++ src/main.cpp
@@ -18,8 +18,8 @@ void seedUniverse(Universe* universe, const std::vector<std::pair<int, int>>& se
 void visualizeUniverse(Universe* universe, size_t time_steps) {
     GridPainter painter(universe->rowCount(), universe->colCount());
     painter.clear();
-    for (int i = 0; i < time_steps; ++i) {
-        for (const std::pair<int, int>& pos: universe->getAliveCellsPos()) {
+    for (size_t i = 0; i < time_steps; ++i) {
+        for (const std::pair<size_t, size_t>& pos: universe->getAliveCellsPos()) {
             painter.paint(pos.first, pos.second, '*', Color::green);
         }
         universe->advance();
